Separated vtkChartHistogram out into separate source files
